requirements.txt
Flask
gspread
oauth2client
matplotlib
numpy
pandas
passlib
python-dotenv